day4/paper.py: Shallow vs. deep copy of array caused unexpected behavior
I discovered that the 'while eligible...' statement at the end was not properly passing the next_array
variable on but was somehow still working. It kept reassigning the original 'line_array' to
the 'next_array' var instead of the array that was modified the last time the loop came around.
This should've resulted in an infinite loop. Instead the script was acting as if the correctly
modified array was being passed. This behavior was confirmed by replacing "line_array" with "next_array"
in the 'find_rolls' function in the while loop and confirming the result was exactly the same.

Debugging revealed the 'find_rolls' function was modifying the global 'line_array' because the
function was modifying the input in place.

Fixed by adding the 'copy' module and explicitly creating a deep copy of the array before modifying in the function.

Made similar changes to 'clear_rolls' function.

Ultimately, this did not affect the functioning of the script since the result was still correct but
if we ever need to reuse this code, it will cause unexplained behavior.

Likely cause is that when Python calls a function, it passes the parameter as a pointer instead of a
copy. This is unexpected and requires further research to fully understand. Will need to make appropriate
adjustments mentally when writing functions in the future.